+++ src/node/ext/zodb/__init__.py
@@ -1,55 +1,63 @@
-from plumber import plumber
+from node.behaviors import Adopt
+from node.behaviors import AsAttrAccess
+from node.behaviors import DefaultInit
+from node.behaviors import NodeChildValidate
+from node.behaviors import Nodify
+from node.behaviors import Order
+from node.ext.zodb.behaviors import OOBTodictStorage
+from node.ext.zodb.behaviors import PodictStorage
+from node.ext.zodb.behaviors import ZODBAttributes
+from node.ext.zodb.behaviors import ZODBBehavior
+from node.ext.zodb.interfaces import IZODBNode
+from node.ext.zodb.utils import OOBTodict
+from node.ext.zodb.utils import Podict
+from node.ext.zodb.utils import volatile_property
 from persistent import Persistent
-from node.behaviors import (
+from plumber import plumber
+from plumber import plumbing
+
+
+@plumbing(
+    NodeChildValidate,
+    Adopt,
+    DefaultInit,
+    Nodify,
+    PodictStorage)
+class ZODBNodeAttributes(Persistent):
+    allow_non_node_childs = True
+
+
+@plumbing(
+    NodeChildValidate,
     Adopt,
     Order,
     AsAttrAccess,
     DefaultInit,
     Nodify,
-    NodeChildValidate,
-)
-from node.ext.zodb.interfaces import IZODBNode
-from node.ext.zodb.behaviors import (
-    ZODBBehavior,
-    PodictStorage,
-    OOBTodictStorage,
     ZODBAttributes,
-)
-from node.ext.zodb.utils import (
-    Podict,
-    OOBTodict,
-    volatile_property,
-)
-
-
-ZODB_ATTRIBUTE_PARTS = (
-    NodeChildValidate, Adopt, DefaultInit, Nodify
-)
-ZODB_NODE_PARTS = (
-    NodeChildValidate, Adopt, Order, AsAttrAccess,
-    DefaultInit, Nodify, ZODBAttributes
-)
-
-
-class ZODBNodeAttributes(Persistent):
-    __metaclass__ = plumber
-    __plumbing__ = ZODB_ATTRIBUTE_PARTS + (PodictStorage,)
-    allow_non_node_childs = True
-
-
+    PodictStorage)
 class ZODBNode(Persistent):
-    __metaclass__ = plumber
-    __plumbing__ = ZODB_NODE_PARTS + (PodictStorage,)
     attributes_factory = ZODBNodeAttributes
 
 
+@plumbing(
+    NodeChildValidate,
+    Adopt,
+    DefaultInit,
+    Nodify,
+    OOBTodictStorage)
 class OOBTNodeAttributes(Persistent):
-    __metaclass__ = plumber
-    __plumbing__ = ZODB_ATTRIBUTE_PARTS + (OOBTodictStorage,)
     allow_non_node_childs = True
 
 
+@plumbing(
+    NodeChildValidate,
+    Adopt,
+    Order,
+    AsAttrAccess,
+    DefaultInit,
+    Nodify,
+    ZODBAttributes,
+    OOBTodictStorage)
 class OOBTNode(Persistent):
-    __metaclass__ = plumber
-    __plumbing__ = ZODB_NODE_PARTS + (OOBTodictStorage,)
     attributes_factory = OOBTNodeAttributes