+++ src/node/ext/zodb/behaviors.py
@@ -1,35 +1,27 @@
-import copy
-from plumber import (
-    Behavior,
-    default,
-    override,
-    finalize,
-    plumb,
-)
-from persistent import Persistent
-from zope.interface import implementer
-from node.interfaces import (
-    INode,
-    IAttributes,
-    IOrdered,
-)
 from node.behaviors import Storage
-from node.utils import (
-    AttributeAccess,
-    instance_property,
-)
 from node.ext.zodb.interfaces import IZODBNode
-from node.ext.zodb.utils import (
-    Podict,
-    OOBTodict,
-)
+from node.ext.zodb.utils import OOBTodict
+from node.ext.zodb.utils import Podict
+from node.interfaces import IAttributes
+from node.interfaces import INode
+from node.interfaces import IOrdered
+from node.utils import AttributeAccess
+from node.utils import instance_property
+from persistent import Persistent
+from plumber import Behavior
+from plumber import default
+from plumber import finalize
+from plumber import override
+from plumber import plumb
+from zope.interface import implementer
+import copy
 
 
 @implementer(IZODBNode)
 class ZODBBehavior(Behavior):
     """This part requires plumbed class to inherit from Persistent.
     """
-    
+
     @override
     @property
     def __parent__(self):
@@ -37,14 +29,14 @@ class ZODBBehavior(Behavior):
         __getitem__.
         """
         return self._v_parent
-    
+
     @override
     def __getitem__(self, key):
         v = self.storage[key]
         if INode.providedBy(v):
             v._v_parent = self
         return v
-    
+
     @override
     def __setattr__(self, name, value):
         """If name is __parent__, write value to _v_parent. This avoids
@@ -54,7 +46,7 @@ class ZODBBehavior(Behavior):
         if name == '__parent__':
             name = '_v_parent'
         Persistent.__setattr__(self, name, value)
-    
+
     @default
     def __call__(self):
         """Meant to be plumbed if something should happen in a particular
@@ -62,12 +54,12 @@ class ZODBBehavior(Behavior):
         mechanism.
         """
         pass
-    
+
     @plumb
     def __setitem__(_next, self, key, value):
         _next(self, key, value)
         self.storage._p_changed = 1
-    
+
     @finalize
     def copy(self):
         return copy.deepcopy(self)
@@ -75,7 +67,7 @@ class ZODBBehavior(Behavior):
 
 @implementer(IOrdered)
 class PodictStorage(ZODBBehavior, Storage):
-    
+
     @default
     @instance_property
     def storage(self):
@@ -84,7 +76,7 @@ class PodictStorage(ZODBBehavior, Storage):
 
 @implementer(IOrdered)
 class OOBTodictStorage(ZODBBehavior, Storage):
-    
+
     @default
     @instance_property
     def storage(self):
@@ -102,8 +94,10 @@ class ZODBAttributes(Behavior):
         try:
             attrs = self._attrs
         except AttributeError:
-            self._attrs = attrs = self.attributes_factory(name='_attrs',
-                                                          parent=self)
+            self._attrs = attrs = self.attributes_factory(
+                name='_attrs',
+                parent=self
+            )
         if self.attribute_access_for_attrs:
             return AttributeAccess(attrs)
         return attrs